p004_ai_nexelin/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.4.0
celery-types==0.22.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.0.9
django-cors-headers==4.5.0
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.3.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.0
docker==7.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langsmith==0.4.35
lxml==6.0.2
numpy==2.3.3
openai==1.54.0
orjson==3.11.3
packaging==25.0
pgvector==0.3.6
Pillow==10.4.0
prompt_toolkit==3.0.52
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.3
qrcode[pil]==7.4.2
redis==5.2.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tqdm==4.67.1
twilio==9.2.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.8.2
zstandard==0.25.0
